Added initial delay to test watch tasks. #4070

+++ .config/bin/trigger-on-stdout-change.js
@@ -8,8 +8,13 @@ var CMD = process.argv[2];
 var CMD_TO_TRIGGER = process.argv[3];
 
 var IS_WINDOWS = /^win/.test(process.platform);
+var waitForFirstBuild = true;
 var triggeredChild;
 var delayedTrigger;
+
+setTimeout(function() {
+  waitForFirstBuild = false;
+}, 5000);
 
 run(CMD, CMD_TO_TRIGGER);
 
@@ -17,9 +22,9 @@ function run(commands, commandToTrigger, spawnOpts) {
   spawnOpts = spawnOpts || {};
 
   if (IS_WINDOWS) {
-      spawnOpts.detached = false;
+    spawnOpts.detached = false;
   }
-  spawnOpts.env = Object.assign({FORCE_COLOR: supportsColor.level}, process.env)
+  spawnOpts.env = Object.assign({FORCE_COLOR: supportsColor.level}, process.env);
 
   var child;
   try {
@@ -32,6 +37,9 @@ function run(commands, commandToTrigger, spawnOpts) {
 
   if (commandToTrigger) {
     child.stdout.on('data', function(data) {
+      if (waitForFirstBuild) {
+        return;
+      }
       if (triggeredChild) {
         treeKill(triggeredChild.pid);
       }
